refactor: make the truncation pairing total in the truncate module
The truncation appliers move from sync.rs into sync/truncate.rs beside
the pure planning rules, and the bare primitives become private to that
module. The exposed surface names each legal composition: the paired
targeted_truncate_wallet_height rewinds range accounting with the data,
for every flow in which the claimed height itself is wrong; the
data-only truncate_data_for_verification exists solely for reorg
fork-finding, where the caller re-prioritises the affected span for
verification and range coverage must survive; and clear_wallet_data is
the clear-all recovery. An unpaired data truncation can no longer be
written by accident, which is the composition hole that let the
reopening path leave condemned testimony seam-eligible.
